use information memory to store time zone and longitude for sidereal time
- when information memory driver is selected in configuration, the sidereal
  time clock uses it
- time zone and longitude are stored when they are being set
- on start the values from information memory are used instead of the
  default ones if available

+++ logic/clock.h
@@ -79,7 +79,7 @@ struct time
 	u32 	last_activity;
 	#ifdef CONFIG_SIDEREAL
 	// offset of local time from UTC (=1: set time is UTC+1 =CET)
-	int		UTCoffset;
+	s8		UTCoffset;
 	#endif
 };
 extern struct time sTime;
